producer/requirements.txt
python-telegram-bot==21.1.1
requests
python-dotenv